requirements.txt
discord.py>=2.0.0
aiosqlite>=0.17.0
requests>=2.28.0
python-dotenv>=0.21.0
apscheduler>=3.10.0
aiohttp>=3.9.0
qrcode[pil]>=7.4.2
pillow>=10.0.0
